quase final

+++ app/actions/usuario/alterar_usuario.php
@@ -6,38 +6,42 @@ require_once("../../config/conecta.php");
 require_once("../../config/upload_imagem.php");
 
 $nome = $_POST['nome'];
-$foto = $_FILES['foto'];
 $desc = $_POST['descricao'];
 $id = $_SESSION['id_usuario'];
 
-$msg1 = uploadImagem($foto);
+if(isset($_FILES['foto'])){
+    $foto = $_FILES['foto'];
+    $msg1 = uploadImagem($foto);
 
-$user = retornaUser($_SESSION['id_usuario']); //função nossa que retorna um objeto usuário
-if ($user->foto !== 'default_user.jpg') {
-    unlink($diretorio . $user->foto); //remove a foto anterior se ela não for a padrão
+    $user = retornaUser($_SESSION['id_usuario']); //função nossa que retorna um objeto usuário
+    if ($user->foto !== 'default_user.jpg') {
+        unlink($diretorio . $user->foto); //remove a foto anterior se ela não for a padrão
+    }
+        
+    conecta();
+
+    global $mysqli;
+
+    $sql = "UPDATE usuario SET foto = ? WHERE idusuario = ?";
+    $stmt = $mysqli->prepare($sql);
+
+    if (!$stmt) {
+        desconecta();
+        return "Erro ao preparar a consulta: " . $mysqli->error;
+    }
+
+    $stmt->bind_param("si", $msg1, $user->idusuario);
+    $stmt->execute();
+
+    if ($stmt->affected_rows > 0) {
+        $msg = "Foto atualizada com sucesso.";
+    } else {
+        $msg = "Não foi possível atualizar a foto.";
+    }
+    desconecta();
 }
 
 conecta();
-
-global $mysqli;
-
-$sql = "UPDATE usuario SET foto = ? WHERE idusuario = ?";
-$stmt = $mysqli->prepare($sql);
-
-if (!$stmt) {
-    desconecta();
-    return "Erro ao preparar a consulta: " . $mysqli->error;
-}
-
-$stmt->bind_param("si", $msg1, $user->idusuario);
-$stmt->execute();
-
-if ($stmt->affected_rows > 0) {
-    $msg = "Foto atualizada com sucesso.";
-} else {
-    $msg = "Não foi possível atualizar a foto.";
-}
-
 
 $sql = "UPDATE usuario SET nome = (?), descricao = (?) WHERE idusuario = (?)";
